requirements.txt
libpysal
matplotlib
networkx
numpy
pyevtk
pynndescent
scikit_learn
scipy
sympy
torchmetrics
tqdm
pandas
openpyxl
pyevtk
libpysal
